Adds failing test comparing two COO matrices that should be the same
The matrix is taken from a program used to calculate the torsion constant of a square beam section.  The 'duplicate' matrix is a direct output from the program.  The 'no duplicates' matrix is the result after passing it through the Scipy sparse matrix routines.  If the order and duplicate normalizing functions are working correctly, these two matrices should be identical.