Fixes NormalizeDuplicates (for sparse COO)
When values are removed, `RowPointers` is modified so that the start of the next row is pointing to the correct element.  Howevever, since we are doing things in-place, when we start reading the next row we sart from the modified location instead of the original location.  To fix this, `index` is set to the end of the original (unmodified) row instead of reading directly from `RowPointers`.
The logic of the internal `while` loop is also simplified.